[1] moves config files and adds logout endpoint
