+++ src/main/resources/application-prod.yml
@@ -1,0 +1,20 @@
+spring:
+  profiles:
+    active: prod
+  datasource:
+    url: jdbc:postgresql://localhost:6432/serp
+    username: serpant
+    password: serpent_dev
+  liquibase:
+    enabled: true
+    change-log: classpath:db/changelog/master.yaml
+    contexts: prod
+    url: jdbc:postgresql://localhost:6432/serp
+  jpa:
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+    hibernate:
+      ddl-auto: validate
+  session:
+    store-type: jdbc
